use svg in bundle and avoid use of img src attribute
loading svgs in img src means that an internet connection is required to load the src if its a url.  Loading svg as a base64 encoded image would take up more space (the raw svg + the base64 encoded one) and require the additional img tag.  Loading as a raw svg component means no img tag or base64 encoding is required.